# Conflicts:
#	src/gameplayer/view/TLGCSValueFinder.java

+++ src/gui/view/TextFieldActorHealthEditor.java
@@ -11,7 +11,7 @@ public class TextFieldActorHealthEditor extends TextFieldWithButton {
 	public TextFieldActorHealthEditor(String labelText, String promptText, Double textFieldWidth) {
 		super(labelText, promptText, textFieldWidth);
 		setButtonAction(e -> {
-			((Actor) getEditableElement()).addAttribute(new Attribute(AttributeType.HEALTH, Integer.parseInt(getTextFieldInput()), (IGameElement) getEditableElement()));
+			((Actor) getEditableElement()).addAttribute(new Attribute(AttributeType.HEALTH, (int) Double.parseDouble((getTextFieldInput())), (IGameElement) getEditableElement()));
 		});
 	}
